backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
lxml==5.2.1
motor==3.4.0
packaging==24.1
passlib==1.7.4
pdfminer.six==20231228
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.7.2
pytest==8.2.2
pytest-asyncio==0.23.6
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
reportlab==4.2.0
requests==2.32.3
sniffio==1.3.0
starlette==0.36.3
typing-inspection==0.4.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.29.0